+++ CHANGELOG.md
@@ -7,4 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Added
+
+- Add `destroy()` function returning the bus.
+
 [Unreleased]: https://github.com/JanekGraff/tmp1075-rs/compare/v0.1.0...HEAD